[ghstack-poisoned]

+++ aten/src/ATen/ParallelNative.cpp
@@ -28,7 +28,7 @@ thread_local bool in_parallel_region_ = false;
 
 // thread number (task_id) set by parallel primitive
 // NOLINTNEXTLINE(cppcoreguidelines-avoid-non-const-global-variables)
-thread_local size_t thread_num_ = 0;
+thread_local int thread_num_ = 0;
 
 void _set_in_parallel_region(bool in_region) {
   in_parallel_region_ = in_region;
@@ -37,7 +37,7 @@ void _set_in_parallel_region(bool in_region) {
 }  // namespace (anonymous)
 
 namespace internal {
-void set_thread_num(size_t thread_num) {
+void set_thread_num(int thread_num) {
   thread_num_ = thread_num;
 }
 }
@@ -109,7 +109,7 @@ void _run_with_pool(const std::function<void(int, size_t)>& fn, size_t range) {
 
 // RAII guard helps to support in_parallel_region() and get_thread_num() API.
 struct ParallelRegionGuard {
-  ParallelRegionGuard(int64_t task_id) {
+  ParallelRegionGuard(int task_id) {
     internal::set_thread_num(task_id);
     _set_in_parallel_region(true);
   }
